+++ Chromosome.h
@@ -14,13 +14,13 @@
 class GeneticAlgo;
 class Chromosome {
 private:
-    std::array<float, CHROMOSOME_LEN> m_genes;
+    std::array<double, CHROMOSOME_LEN> m_genes;
     GeneticAlgo* m_ga;
     double m_fitness;
     
 public:
     Chromosome();
-    Chromosome(std::array<float, CHROMOSOME_LEN>&, double fitness=0.0);
+    Chromosome(std::array<double, CHROMOSOME_LEN>&, double fitness=0.0);
     void setFitness(double);
     double getFitness();
     void getRandomBits();
@@ -28,12 +28,12 @@ public:
     void printDecoded();
     void setGeneticAlgo(GeneticAlgo* ga);
     void calcFitness();
-    static void mutate(std::array<float, CHROMOSOME_LEN>& child);
-    static void crossover(std::array<float, CHROMOSOME_LEN>& child1, std::array<float, CHROMOSOME_LEN>& child2);
-    static std::array<float, CHROMOSOME_LEN>& rouletteSelect(double totalFitness, Chromosome c_arr[], int len);
+    static void mutate(std::array<double, CHROMOSOME_LEN>& child);
+    static void crossover(std::array<double, CHROMOSOME_LEN>& child1, std::array<double, CHROMOSOME_LEN>& child2);
+    static std::array<double, CHROMOSOME_LEN>& rouletteSelect(double totalFitness, Chromosome c_arr[], int len);
     void decode();
     void to_vector(std::vector<std::vector<double>>& vec);
-    static void prune(std::array<float, CHROMOSOME_LEN>& child);
+    static void prune(std::array<double, CHROMOSOME_LEN>& child);
 
     
 };
